Codeforces

+++ Codeforces/546B.cpp
@@ -20,7 +20,7 @@ using namespace std;
 void solve()
 {
        ll n; cin >> n;
-       vector<int> a(n);
+       vector<ll> a(n);
        for(int i = 0; i < n; i++){
               cin >> a[i];
        }
